fix: 修复视频上传功能
- 修复文件保存逻辑，将文件保存从 Controller 层移到 Service 层
- 修复登录拦截器，支持 Bearer Token 格式
- 修复 RabbitMQ 消息序列化问题，使用 JSON 转换器
- 移除视频上传 API 的限流
- 配置日志编码为 UTF-8
- 修改封面文件存储路径到 /static/coverUrl
- 优化前端上传页面，删除按钮一直显示红色背景

+++ src/main/java/com/yuliyuli/document/VideoDocument.java
@@ -22,7 +22,7 @@ public class VideoDocument {
   private String title;
 
   // 作者Id，用于点击作者名字后进行作者主页跳转
-  @Field(type = FieldType.Integer)
+  @Field(type = FieldType.Long)
   private Long userId;
 
   @Field(type = FieldType.Text)
@@ -40,8 +40,8 @@ public class VideoDocument {
   @Field(type = FieldType.Integer)
   private Integer typeId;
 
-  @Field(type = FieldType.Integer)
-  private Integer playCount;
+  @Field(type = FieldType.Long)
+  private Long playCount;
 
   @Field(type = FieldType.Date)
   private Date createTime;
